refactor(catalog): organize folders
