+++ src/main/java/me/helioalbano/biblioteca/catalog/usecase/addnewbook/AddNewBook.java
@@ -1,5 +1,7 @@
 package me.helioalbano.biblioteca.catalog.usecase.addnewbook;
 
+import me.helioalbano.biblioteca.catalog.usecase.addnewbook.dto.AddNewBookInput;
+
 public interface AddNewBook {
-    Output execute(Input input);
+    void execute(AddNewBookInput addNewBookInput);
 }
