--- src/main/java/me/helioalbano/biblioteca/catalog/usecase/addnewbook/Input.java
@@ -1,13 +1,0 @@
-package me.helioalbano.biblioteca.catalog.usecase.addnewbook;
-
-public class Input {
-    private String title;
-
-    public Input(String title) {
-        this.title = title;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-}
