+++ src/main/java/me/helioalbano/biblioteca/catalog/usecase/addnewbook/impl/AddNewBookImpl.java
@@ -1,9 +1,10 @@
-package me.helioalbano.biblioteca.catalog.usecase.addnewbook;
+package me.helioalbano.biblioteca.catalog.usecase.addnewbook.impl;
 
 import me.helioalbano.biblioteca.catalog.domain.entity.Book;
 import me.helioalbano.biblioteca.catalog.domain.repository.BookRepository;
 import me.helioalbano.biblioteca.catalog.domain.valueobject.Title;
-import me.helioalbano.biblioteca.catalog.usecase.addnewbook.dto.Input;
+import me.helioalbano.biblioteca.catalog.usecase.addnewbook.AddNewBook;
+import me.helioalbano.biblioteca.catalog.usecase.addnewbook.dto.AddNewBookInput;
 
 public class AddNewBookImpl implements AddNewBook {
 
@@ -13,8 +14,8 @@ public class AddNewBookImpl implements AddNewBook {
         this.repository = repository;
     }
 
-    public void execute(Input input) {
-        Book book = new Book(new Title(input.getTitle()));
+    public void execute(AddNewBookInput addNewBookInput) {
+        Book book = new Book(new Title(addNewBookInput.getTitle()));
         repository.create(book);
     }
