+++ src/test/java/me/helioalbano/biblioteca/catalog/usecase/addnewbook/AddNewBookTest.java
@@ -1,8 +1,11 @@
 package me.helioalbano.biblioteca.catalog.usecase.addnewbook;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.*;
 
+import me.helioalbano.biblioteca.catalog.domain.entity.Book;
+import me.helioalbano.biblioteca.catalog.usecase.addnewbook.dto.AddNewBookInput;
+import me.helioalbano.biblioteca.catalog.usecase.addnewbook.impl.AddNewBookImpl;
 import org.junit.jupiter.api.Test;
 
 import me.helioalbano.biblioteca.catalog.domain.repository.BookRepository;
@@ -11,9 +14,8 @@ public class AddNewBookTest {
     @Test
     void addAnInexistentBook() {
         BookRepository repository = mock(BookRepository.class);
-        AddNewBook addNewBook = new AddNewBookImpl(repository);
-        Input input = new Input("O Programador Pragmático");
-        Output output = addNewBook.execute(input);
-        assertEquals("O Programador Pragmático", output.getTitle());
+        var input = new AddNewBookInput("O Programador Pragmático");
+        new AddNewBookImpl(repository).execute(input);
+        verify(repository, times(1)).create(any(Book.class));
     }
 }
